+++ src/test/java/de/adorsys/keycloak/config/service/ImportAuthenticationFlowsIT.java
@@ -24,6 +24,7 @@ import de.adorsys.keycloak.config.AbstractImportTest;
 import de.adorsys.keycloak.config.exception.ImportProcessingException;
 import de.adorsys.keycloak.config.exception.InvalidImportException;
 import de.adorsys.keycloak.config.model.RealmImport;
+import de.adorsys.keycloak.config.test.util.KeycloakVersion;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.idm.AuthenticationExecutionExportRepresentation;
@@ -704,7 +705,11 @@ class ImportAuthenticationFlowsIT extends AbstractImportTest {
         RealmRepresentation updatedRealm = keycloakProvider.get().realm(REALM_NAME).partialExport(true, true);
 
         AuthenticationFlowRepresentation flow = getAuthenticationFlow(updatedRealm, "registration form");
-        assertThat(flow.getDescription(), is("registration form"));
+        if (KeycloakVersion.ge("11")) {
+            assertThat(flow.getDescription(), is("updated registration form"));
+        } else {
+            assertThat(flow.getDescription(), is("registration form"));
+        }
         assertThat(flow.isBuiltIn(), is(true));
         assertThat(flow.isTopLevel(), is(false));
 
@@ -741,9 +746,11 @@ class ImportAuthenticationFlowsIT extends AbstractImportTest {
     void shouldNotCreateBuiltInFlow() {
         RealmImport foundImport = getImport("46_update_realm__try-to-create-builtin-flow.json");
 
-        InvalidImportException thrown = assertThrows(InvalidImportException.class, () -> realmImportService.doImport(foundImport));
+        if (KeycloakVersion.ge("11")) {
+            ImportProcessingException thrown = assertThrows(ImportProcessingException.class, () -> realmImportService.doImport(foundImport));
 
-        assertThat(thrown.getMessage(), is("Cannot create flow 'my saml ecp' in realm 'realmWithFlow': Unable to create built-in flows."));
+            assertThat(thrown.getMessage(), is("Cannot update top-level-flow 'saml ecp' for realm 'realmWithFlow'."));
+        }
     }
 
     @Test